Move blockMemHashOptions from Server to Stats
so that when Envoy compiles in not-hot-restart mode and elides the hot_restart_impl files, we can still use the relevant utilities
