+++ test/common/stats/stats_impl_test.cc
@@ -8,7 +8,6 @@
 #include "common/config/well_known_names.h"
 #include "common/stats/stats_impl.h"
 
-#include "server/hot_restart_impl.h"
 #include "server/options_impl.h"
 
 #include "test/mocks/server/mocks.h"
@@ -479,7 +478,7 @@ public:
   void setup() {
     Stats::RawStatData::configureForTestsOnly(options_);
     BlockMemoryHashSetOptions memory_hash_set_options_ =
-        Server::blockMemHashOptions(options_.maxStats());
+        Stats::blockMemHashOptions(options_.maxStats());
     uint32_t num_bytes = BlockMemoryHashSet<Stats::RawStatData>::numBytes(memory_hash_set_options_);
     memory_.reset(new uint8_t[num_bytes]);
     memset(memory_.get(), 0, num_bytes);
